requirements.txt
requests
python-dotenv
pandas
pytest
yfinance
tqdm
